feat: Add requestUpdateStatusList function to API
This commit adds the `requestUpdateStatusList` function to the API. This function sends a PUT request to update the status list with the provided `statusList` data. If the response is not successful, an error is thrown with a corresponding error message.

The following files were modified:
- src/api/status.js: Added `requestUpdateStatusList` function
- src/components/ProjectSettingsDialog.jsx: Updated import statements for hooks and added `useUpdateStatusList` hook
- src/components/ProjectSettingsDialog.jsx: Replaced call to `updateStatus` mutation with `updateStatusList`
- src/components/ProjectSettingsDialog.jsx: Updated logic in `handleDragEnd` function to create an updated status list and call `updateStatusList`
- src/components/StatusCard.jsx: Uncommented key prop in Card component
- src/hooks/status.js: Added `requestUpdateStatusList` import and exported new hook `useUpdateStatusList`

Note that console.log statement was added for debugging purposes and will be removed before merging this code.