feat(ToDoContext): update initial status for todoFormDialog
This commit adds the ability to set the initial status for new tasks in the ToDoContext. The code now retrieves the list of statuses and sets the first status as the default value for new tasks. This ensures that when a user creates a new task, it will have an appropriate initial status.