feat(app): Add useMediaQuery hook and isMobile state
This commit adds the `useMediaQuery` hook from the Material-UI library to the `AppContext.jsx` file. It also introduces a new state variable called `isMobile`, which is set based on the screen width using a media query. This change allows for responsive behavior in the application.

fix(layout): Remove unnecessary query invalidations

This commit removes unnecessary query invalidations in the `Layout.jsx` file. The commented out lines that invalidate queries related to user, status list, and to-do list have been removed as they are no longer needed.